add Gen3, more performant Gen2
Gen3 refactors Gen2 by moving the template parsing step at
initialization, instead of doing it in the Emit function.

To achieve this, GenState and dupes are **always** set to an
empty value. Is not clear to me which impact this have on
generated data.

Gen3 is still allocation and memory heavy but way less than Gen2.

Benchmark:
name                          time/op
_Gen2-12                      64.8µs ± 9%
_Gen3-12                      15.3µs ± 7%
_Generator-12                 39.9µs ± 6%
_GeneratorWithTemplate-12     39.3µs ± 8%
_GeneratorWithTemplateAWS-12  8.74µs ± 7%

name                          alloc/op
_Gen2-12                      10.8kB ± 0%
_Gen3-12                      2.94kB ± 0%
_Generator-12                   482B ± 0%
_GeneratorWithTemplate-12       482B ± 0%
_GeneratorWithTemplateAWS-12    160B ± 0%

name                          allocs/op
_Gen2-12                         219 ± 0%
_Gen3-12                        97.0 ± 0%
_Generator-12                   15.0 ± 0%
_GeneratorWithTemplate-12       15.2 ± 5%
_GeneratorWithTemplateAWS-12    5.00 ± 0%