output string on missing field
In case `generate()` is being used in the template on a missing field
the current behaviour of returning an empty string hides the problem to
the template developer.

As it is not expected for the template to try generating a missing
field, this commit returns a known string highlighting the problem.