Fix the flakiness of cardinality tests
For each cardinality being tested, a random range of values is chosen.
However if such range is too small for the asked cardinality, the test fails.

This change adds a retry logic to discard too small ranges.

The code is organized to later support cardinality based on the field type.